+++ src/ArgsAwareTrait.php
@@ -2,10 +2,9 @@
 
 namespace Dhii\Invocation;
 
-use Exception as RootException;
 use InvalidArgumentException;
+use stdClass;
 use Traversable;
-use Dhii\Util\String\StringableInterface as Stringable;
 
 /**
  * Functionality for args awareness.
@@ -19,7 +18,7 @@ trait ArgsAwareTrait
      *
      * @since [*next-version*]
      *
-     * @var array|Traversable
+     * @var array|Traversable|stdClass|null
      */
     protected $args;
 
@@ -28,11 +27,13 @@ trait ArgsAwareTrait
      *
      * @since [*next-version*]
      *
-     * @return array|Traversable The args.
+     * @return array|Traversable|stdClass The args.
      */
     protected function _getArgs()
     {
-        return $this->args;
+        return !is_null($this->args)
+            ? $this->args
+            : array();
     }
 
     /**
@@ -40,49 +41,21 @@ trait ArgsAwareTrait
      *
      * @since [*next-version*]
      *
-     * @param array|Traversable $args A list of argument values.
+     * @param array|Traversable|stdClass $args A list of argument values.
      *
      * @throws InvalidArgumentException If the arguments list is invalid.
      */
     protected function _setArgs($args)
     {
-        if (!is_array($args) && !($args instanceof Traversable)) {
-            throw $this->_createInvalidArgumentException($this->__('Invalid args list'), null, null, $args);
-        }
+        $args = $this->_normalizeIterable($args);
 
         $this->args = $args;
     }
 
     /**
-     * Creates a new invalid argument exception.
-     *
-     * @since [*next-version*]
-     *
-     * @param string|Stringable|null $message  The error message, if any.
-     * @param int|null               $code     The error code, if any.
-     * @param RootException|null     $previous The inner exception for chaining, if any.
-     * @param mixed|null             $argument The invalid argument, if any.
-     *
-     * @return InvalidArgumentException The new exception.
-     */
-    abstract protected function _createInvalidArgumentException(
-        $message = null,
-        $code = null,
-        RootException $previous = null,
-        $argument = null
-    );
-
-    /**
      * Translates a string, and replaces placeholders.
      *
-     * @since [*next-version*]
-     * @see   sprintf()
-     *
-     * @param string $string  The format string to translate.
-     * @param array  $args    Placeholder values to replace in the string.
-     * @param mixed  $context The context for translation.
-     *
-     * @return string The translated string.
+     * @return array|Traversable|stdClass The normalized iterable.
      */
-    abstract protected function __($string, $args = array(), $context = null);
+    abstract protected function _normalizeIterable($iterable);
 }