+++ CHANGELOG.md
@@ -5,4 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Fixed
+- Exception type throw by `InvokeCallableCapableTrait#_invokeCallable()` is now properly documented to be
+`InvocationExceptionInterface` instead of `RootException`.
+
+### Changed
+- `_setArgs()` and `_invokeCallback()` now normalize the args list, instead of manually validating it.
+- `_setArgs()`, `_invokeCallback()`, and `_invokeCallable()` now accept `stdClass` as arg list.
+
+## [0.1-alpha1] - 2018-02-12
 Initial version.